IT

Description

An IT expression represents a function of the form:

\[ f(x) = \sum_{i}{w_i \cdot t_i(\prod_{j}{x_j^{k_{ij}})}} \]

with \(t_i\) being a transformation function.

Any given expression can be represented by a list of terms, with each term being composed of a transformatioon function and an interaction. The transformation function is represented by a String of its name and a function of the type `a -> a`. The interaction is represented as a `Map Int Int` where the key is the predictor index and the value is the strength of the predictor in this term. The strengths with a value of zero are omitted.

Documentation

type Interaction = Map Int Int

The Interaction type is a map where a key, value pair (i,p) indicates that the i-th variable should be raised to the power of p.

data Transformation a

The Transformation type contains the name of the function and the transformation function.

Constructors

Transformation String (a -> a)

data Term a

A Term is the product type of a Transformation and an Interaction.

Constructors

Term (Transformation a) Interaction

newtype Expr a

An Expr is just a list of Terms.

Constructors

Expr [Term a]

type Column a = Vector a

A Column of a data set is stored as a LA.Vector to avoid conversions during the fitting of the coefficients.

type Dataset a = Vector (Column a)

The Dataset is a Vector of Columns for efficiency. TODO: use Accelerate for large data sets.

Class IT of IT-expressions

class IT a where

The IT class defines how we should evaluate an expression given the type a.

Methods

itTimes :: Dataset a -> Interaction -> Column a

itTimes evaluates the interaction of an expression using the values in the Dataset, it returns a single Column representing the new predictor.

itAdd :: [Column a] -> Column a

itAdd sums a list of terms.

itWeight :: Double -> Column a -> Column a

itWeight applies the coefficients to a Column.

evalTerm :: (Storable a, Container Vector a, IT a) => Term a -> Dataset a -> Column a

to evaluate a term we apply the transformation function to the result of itTimes.

evalExprToList :: (Storable a, Container Vector a, IT a) => Expr a -> Dataset a -> [Column a]

evaluates the expression into a list of terms in case the evaluated values are needed

evalExpr :: (Storable a, Container Vector a, IT a) => Expr a -> Dataset a -> [Double] -> Column a

evaluates an expression by evaluating the terms into a list applying the weight and summing the results.

Utility functions

uniqueTerms :: Expr a -> Expr a

Remove duplicated terms.

hasTerm :: Expr a -> Term a -> Bool

Check whether and expression has a given term.

Internal functions

removeIthTerm :: Int -> Expr a -> Expr a

remove the i-th term of an expression.

getIthTerm :: Int -> Expr a -> Maybe (Term a)

returns the i-th term of an expression.

consTerm :: Term a -> Expr a -> Expr a

inserts a new term into the expression.

getListOfTerms :: Expr a -> [Term a]

get the list of terms

numberOfTerms :: Expr a -> Int

returns the number of terms in an expression.

getInteractions :: Term a -> Interaction

return the interactions of a term