IT.Algorithms

Description

Definitions for a Solution, Population of solutions, fitness function and mutation function.

Documentation

data Solution a

data type containing a solution, its fitness and weight vector a refers to the type of Expr, b refers to a container of statistics.

Constructors

Sol

Fields

_expr :: Expr a

The IT expression of type a

_fit :: [Double]

Fitness and other measures for evaluating the expression

_weights :: [Vector]

type Population a = [Solution a]

A population of 'Solution a b'

type Fitness a = Expr a -> Maybe (Solution a)

Fitness function that takes a list of expressions and returns an evaluated population. This function is a good candidate for parallelization. type Fitness a b = [Expr a] -> Population a b -- (Expr a, Double, b)

type Mutation a = Expr a -> Rnd (Expr a)

Mutation function with signature 'Solution a b -> Rnd (Solution a b)'