IT.Eval

Description

Definitions of IT data structure and support functions.

Documentation

regId :: Transformation Double

Transformation Functions

regSin :: Transformation Double

Transformation Functions

regCos :: Transformation Double

Transformation Functions

regTan :: Transformation Double

Transformation Functions

regTanh :: Transformation Double

Transformation Functions

regSqrt :: Transformation Double

Transformation Functions

regAbsSqrt :: Transformation Double

Transformation Functions

regLog :: Transformation Double

Transformation Functions

regExp :: Transformation Double

Transformation Functions

regAll :: [Transformation Double]

List of all transformation functions

regTrig :: [Transformation Double]

regNonLinear :: [Transformation Double]

regLinear :: [Transformation Double]

toTrans :: String -> Transformation Double

Convert a string to a Transformation Function

isInvalid :: Double -> Bool

A value is invalid if it's wether NaN or Infinite

isValid :: [Double] -> Bool

a set of points is valid if none of its values are invalid and the maximum abosolute value is below 1e150 (to avoid overflow)

exprToMatrix :: Dataset Double -> Expr Double -> Matrix Double

evaluate an expression to a set of samples

(1 LA.|||) adds a bias dimension

cleanExpr :: Dataset Double -> Expr Double -> Expr Double

Clean the expression by removing the invalid teerms

notInfNan :: Solution Double -> Bool

Checks if the fitness of a solution is not Inf nor NaN.

Orphan instances

IT Double
IT Instance for regression evals an expression to sum(w_i * term_i) + w0 with term_i = f_i(p_i(xs)) and p_i(xs) = prod(x^eij)

Instance details

Methods

itTimes :: Dataset Double -> Interaction -> Column Double

itAdd :: [Column Double] -> Column Double

itWeight :: Double -> Column Double -> Column Double