IT.ITEA

Description

Generic implementation of Interaction-Transformation Evolutionary Algorithm for any instance of IT expression.

To run itea you just need to call 'itea mutFun pop0', where mutFun is a mutation function of the type 'Mutation a b', and pop0 is the initial 'Population a b' of solutions. This function will result in an infinite list of populations, with the i-th element being the population of the i-th generation.

This library also provides some generic mutation function builders.

A sample example is provided by ITEAReg module.

ITEA

itea :: (NFData a, NFData b) => Mutation a -> Fitness a b -> Population a b -> Rnd [Population a b]

Creates an infinite list of populations where the i-th element corresponds to t he i-th generation.

initialPop :: Int -> Int -> Int -> Rnd (Term a) -> Fitness a b -> Rnd (Population a b)

Generate an Initial Population at Random

tournament :: Population a b -> Int -> Rnd (Population a b)

Tournament Selection

given the concatenation of the previous population and the mutated children, it will return a sampled selection of these combined population with the same size as the original population.

step :: (NFData a, NFData b) => Mutation a -> Fitness a b -> Int -> Population a b -> Rnd (Population a b)

Perform one iterative step of ITEA