IT.Metrics

Description

Definitions of support functions to calculate a set of error measures for regression and classification.

Documentation

type Vector = Vector Double

Performance measures

data Measure

Constructors

Measure

Fields

_name :: String

_fun :: Vector -> Vector -> Double

true values -> predicted values -> measure

mean :: Vector -> Double

Mean for a vector of doubles

var :: Vector -> Double

Variance for a vector of doubles

meanError

Arguments

:: (Vector -> Vector)
a function to be applied to the error terms (abs, square,...)

-> Vector
fitted values

-> Vector
target values

-> Double

generic mean error measure

Common error measures for regression:

mse :: Vector -> Vector -> Double

Mean Squared Error

mae :: Vector -> Vector -> Double

Mean Absolute Error

nmse :: Vector -> Vector -> Double

Normalized Mean Squared Error

rmse :: Vector -> Vector -> Double

Root of the Mean Squared Error

rSq :: Vector -> Vector -> Double

negate R^2 - minimization metric

Regression measures

_rmse :: Measure

_mae :: Measure

_nmse :: Measure

_r2 :: Measure

Classification measures

_accuracy :: Measure

_recall :: Measure

_precision :: Measure

_f1 :: Measure

_logloss :: Measure

accuracy :: Vector -> Vector -> Double

Accuracy: ratio of correct classification

precision :: Vector -> Vector -> Double

Precision: ratio of correct positive classification

recall :: Vector -> Vector -> Double

Recall: ratio of retrieval of positive labels

f1 :: Vector -> Vector -> Double

Harmonic average between Precision and Recall

logloss :: Vector -> Vector -> Double

LogLoss of a classifier that returns a probability.

measureAll :: [Measure]

List of all measures

toMeasure :: String -> Measure

Read a string into a measure