IT.Random

Description

Random expressions generation

type Rnd a = State StdGen a

The type used for generating random values of a

sampleInterMax

Arguments

:: Int
problem dimension

-> Int
maximum number of variables in this interaction

-> Int
minimum exponent

-> Int
maximum exponent

-> Rnd Interaction
Random interaction generator

Creates a random interaction with up to n variables. Guaranteed to choose at least one variable.

sampleInterMax' :: Int -> Int -> Int -> Int -> Rnd [(Int, Int)]

sampleInter

Arguments

:: Int
problem dimension

-> Int
minimum exponent

-> Int
maximum exponent

-> Rnd Interaction
Random interaction generator

Sample a random interaction given the problem dimension and the minimum and maximum exponents

sampleInter' :: Int -> Int -> Int -> Rnd [(Int, Int)]

sampleTrans

Arguments

:: [Transformation]
choices of transformation functions

-> Rnd Transformation
Random generator

Samples a random transformation function from a provided list of functions

sampleTerm

Arguments

:: Rnd Transformation
random transformation function

-> Rnd Interaction
random interaction function

-> Rnd Term
Random generator

Samples a random term using a random transformation and a random interaction generators.

sampleExpr

Arguments

:: Rnd Term
random term function

-> Int
number of terms

-> Rnd Expr
Random generator

Create a random expression with exactly n terms

samplePop :: Int -> Int -> (Int -> Rnd Expr) -> Rnd [Expr]

Create a random population of n expressions with varying number of terms

Utility random functions

sampleTo :: Int -> Rnd Int

sample from [0,n]

sampleRng :: Int -> Int -> Rnd Int

Sample from a range of integers

sampleNZRng :: Int -> Int -> Rnd Int

Sample from a range of integers excluding 0

sampleFromList :: [a] -> Rnd a

Sample a random element from a list

toss :: Rnd Bool

Toss a coin to your Witcher