IT.Regression

Description

Definitions of IT data structure and support functions.

IT specific stuff

data Task

Constructors

Regression

Classification

ClassMult

type FitFun = Vector -> Vector -> Double

predict :: Matrix Double -> Vector -> Vector

Predict a linear model

solveOLS :: Matrix Double -> Vector -> Vector

Solve the OLS *zss*w = ys*

regress :: Vector -> Matrix Double -> Maybe (Vector, [Vector])

Applies OLS and returns a Solution if the expression is invalid, it returns Infinity as a fitness

classify :: Vector -> Matrix Double -> Maybe (Vector, [Vector])

classifyMult :: Vector -> Matrix Double -> Maybe (Vector, [Vector])

evalTrain :: Task -> NonEmpty Measure -> Dataset Double -> Vector -> Expr Double -> Maybe (Solution Double)

Fitness function for regression

Split the dataset into twice the available cores evaluate the expressions in parallel run a Linear regression on the evaluated expressions Remove from the population any expression that leads to NaNs or Infs it was fitnessReg

evalTest :: Task -> NonEmpty Measure -> Dataset Double -> Vector -> Solution Double -> Maybe [Double]

Evaluates an expression into the test set. This is different from fitnessReg since it doesn't apply OLS. It was: fitnessTest

roundoff :: RealFrac a => a -> a

Experimental: round off floating point to the 1e-10 place.

tryToRound :: (Vector -> Double) -> Matrix Double -> (Vector, [Vector]) -> (Vector, [Vector])