ITEA.Config

Documentation

class Monoid a => Valid a b | a -> b, b -> a where

Configuration Validation

Methods

validateConfig :: a -> b

data Param a

Constructors

None

Has a

fromParam :: Param a -> a

data Funcs

Mutation configuration

Constructors

FLinear

FNonLinear

FTrig

FAll

data UncheckedMutationCfg

Constructors

UMCfg

Fields

_expLim :: Param (Int, Int)

_termLim :: Param (Int, Int)

_nzExp :: Param Int

_transFun :: Param Funcs

data MutationCfg

Constructors

MCfg (Int, Int) (Int, Int) Int Funcs

exponents :: Int -> Int -> UncheckedMutationCfg

termLimit :: Int -> Int -> UncheckedMutationCfg

nonzeroExps :: Int -> UncheckedMutationCfg

transFunctions :: Funcs -> UncheckedMutationCfg

getMaxTerms :: MutationCfg -> Int

parseFile :: String -> (Matrix Double, Vector)

Parse a numerical csv file into predictors and target variables

withMutation :: MutationCfg -> Int -> (Mutation Double, Rnd (Term Double))

Creates the mutation function and also returns the random term generator (for initialization)

data UncheckedDatasets

Datasets configuration

Constructors

UD

Fields

_trainset :: Param String

_testset :: Param String

data Datasets

Constructors

D String String

trainingset :: String -> UncheckedDatasets

testset :: String -> UncheckedDatasets

data Output

Output configuration

Constructors

Screen

PartialLog String

FullLog String

getBest :: Int -> [Population Double RegStats] -> Solution Double RegStats

getAllBests :: (Foldable t, Ord b) => Int -> [t b] -> [b]

data StatType

Constructors

Best

Worst

Avg

applyStat :: (a -> b) -> Int -> [[a]] -> [[b]]

getAll :: StatType -> (RegStats -> Double) -> Int -> [[RegStats]] -> [Double]

getAllStats :: Int -> [[RegStats]] -> [AggStats]

gets all stats at once instead of going through the list multiple times

createIfDoesNotExist :: FilePath -> IO Handle

Creates a file if it does not exist

genReports :: Output -> [Population Double RegStats] -> Int -> (Solution Double RegStats -> RegStats) -> IO ()

Generates the reports into the output

data AggStats

Constructors

AS

Fields

_best :: RegStats

_worst :: RegStats

_avg :: RegStats

data Metric

Constructors

RMSE

MAE

NMSE

R2

genEvoStream :: [RegStats] -> [Handle] -> IO ()

genEvoReport :: [AggStats] -> String -> IO ()

resultsToStr :: Solution Double RegStats -> RegStats -> [String]

headReport :: [Char]