ITEA.Report

Description

Configuration parsing and report generation.

Documentation

data Output

Output configuration

Constructors

Screen

PartialLog String

FullLog String

getBest :: Int -> [Population] -> Solution

Get best solution from all generations

getBestMaybe :: Int -> [Population] -> Maybe Solution

getAllBests :: Int -> [Population] -> [Solution]

createIfDoesNotExist :: String -> FilePath -> IO Handle

Creates a file if it does not exist

interleave :: [a] -> [a] -> [a]

genReports :: Output -> NonEmpty Measure -> [Population] -> Int -> (Solution -> Maybe [Double]) -> (Expr -> Maybe Solution) -> IO ()

Generates the reports into the output

openNext :: String -> IO Handle

Opens the first file available in the format "name.{i}.csv" where i follows a sequence from 0 onward.

postAgg :: [Double] -> [Double]

aggregate :: [Double] -> Double -> [Double]

toFile :: [Handle] -> [Double] -> IO ()

Write a sequence to a sequence of opened files