RunConfig

Description

Main program to run ITEA with a configuration file.

Documentation

data Alg

Algorithm Choice

ITEA is the mutation-based Symbolic Regression for IT representation FI2POP splits the population as feasible and infeasible to deal with constraints

Constructors

ITEA

FI2POP

getSetting :: Get_C a => ConfigParser -> SectionSpec -> OptionSpec -> a

Read the option x located at section cat from configuration cp.

getWithDefault :: Get_C a => a -> ConfigParser -> SectionSpec -> OptionSpec -> a

get a setting value if it exists in the config file, or return a default value

runWithConfig :: String -> IO ()

Read the config file and run the algorithm.

parseConfigFile :: [String] -> IO ()

Parse the filename from the system arguments.