RunCrossVal

Description

Run a 5-fold cross-validation on the training set to find the best combination of hyperparameters.

TODO: refactor, document, write the best configuration in a config file

Documentation

createMutCfg :: (Int, Int) -> Int -> [Transformation] -> MutationCfg

Returns the mutation configuration, population size and number of generations.

'(e1, e2)' are the minimum and maximum exponentes.

tmax is the maximum number of terms.

pop is the population size.

validateArgs :: [String] -> (String, Int)

Validates the program arguments

runITEARegCV

Arguments

:: Fitness
Training fitness function

-> (Solution -> Maybe [Double])
Test fitness function

-> Int
Problem dimension

-> MutationCfg
Mutation configuration

-> Int
population size

-> Int
number of generations

-> IO Double

Runs a single experiment with a given configuration

runFI2POPRegCV

Arguments

:: Fitness
Training fitness function

-> (Solution -> Maybe [Double])
Test fitness function

-> Int
Problem dimension

-> MutationCfg
Mutation configuration

-> Int
population size

-> Int
number of generations

-> IO Double

runCfg :: FilePath -> Int -> MutationCfg -> IO Double

runs a configuration for a given data set

runCrossVal :: [String] -> IO ()

Main function