IT

Description

An IT expression represents a function of the form:

\[ f(x) = \sum_{i}{w_i \cdot t_i(\prod_{j}{x_j^{k_{ij}})}} \]

with \(t_i\) being a transformation function.

Any given expression can be represented by a list of terms, with each term being composed of a transformatioon function and an interaction. The transformation function is represented by a Transformation sum type. The interaction is represented as an `IntMap Int` where the key is the predictor index and the value is the strength of the predictor in this term. Strengths with a value of zero are omitted.

Documentation

type Interaction = IntMap Int

The Interaction type is a map where a key, value pair (i,p) indicates that the i-th variable should be raised to the power of p.

data Transformation

The Transformation type describes the function that should be evaluated. The evaluation is defined in Eval module.

Constructors

Id

Sin

Cos

Tan

Tanh

Sqrt

SqrtAbs

Log

Exp

Log1p

data Term

A Term is the product type of a Transformation and an Interaction.

Constructors

Term Transformation Interaction

type Expr = [Term]

An Expr is just a list of Terms.

type Column a = Vector a

A Column of a data set is stored as a LA.Vector to avoid conversions during the fitting of the coefficients.

type Dataset a = Vector (Column a)

The Dataset is a Vector of Columns for efficiency. TODO: try Repa or Accelerate for large data sets.

prettyPrint

Arguments

:: ((Int, Int) -> String)
A function that converts an interaction to a string

-> (Transformation -> String)
A function that converts a transformation to a string

-> Expr
The expression to convert to string

-> [Double]
The fitted coefficients

-> String
A string representing the expression

Base interface for converting an expression to string

toExprStr :: Expr -> [Double] -> String

Converts an expression to a readable format

toPython :: Expr -> [Double] -> String

Converts an expression to a numpy compatible format

Internal functions

removeIthTerm :: Int -> Expr -> Expr

remove the i-th term of an expression.

getIthTerm :: Int -> Expr -> Maybe Term

returns the i-th term of an expression.

getInteractions :: Term -> Interaction

return the interactions of a term

exprLength :: Expr -> [Double] -> Int

returns the length of an expression as in https://github.com/EpistasisLab/regression-benchmark/blob/dev/CONTRIBUTING.md

termLength :: Term -> Int

The length of a term is the interaction length plus 1 if the transformation function is not Id.

interactionLength :: Interaction -> Int

The interaction length is calculated as: +2 for every exponent different from 0 and 1 (^k) +1 for every nonzero exponent, except for the first (*) +1 for every nonzero exponent (x_i)