IT.Algorithms

Description

Definitions for a Solution, Population of solutions, fitness function and mutation function.

Documentation

data Solution

data type containing a solution, its fitness and weight vector a refers to the type of Expr, b refers to a container of statistics.

Constructors

Sol

Fields

_expr :: Expr

The IT expression of type a

_fit :: [Double]

Fitness and other measures for evaluating the expression

_constr :: Double

Amount of Shape Constraint violation associated with the expression, always positive

_len :: Int

Expression size as per https://github.com/EpistasisLab/regression-benchmark/blob/dev/CONTRIBUTING.md

_penalty :: Double

penalty of fitness

_weights :: [Vector]

Weights associated with the expression (they count towards the length)

type Population = [Solution]

A population of 'Solution a b'

type Fitness = Expr -> Maybe Solution

Fitness function that takes a list of expressions and returns an evaluated population. This function is a good candidate for parallelization. type Fitness a b = [Expr a] -> Population a b -- (Expr a, Double, b)

type Constraint = Expr -> [Double] -> Double

Constraint is a function that receives an expression and its coefficients and return the penalty associated with the constraint violation.

type Mutation = Expr -> Rnd Expr

Mutation function with signature 'Solution a b -> Rnd (Solution a b)'