IT.ITEA

Description

Generic implementation of Interaction-Transformation Evolutionary Algorithm for any instance of IT expression.

To run itea you just need to call 'itea mutFun pop0', where mutFun is a mutation function of the type 'Mutation a b', and pop0 is the initial 'Population a b' of solutions. This function will result in an infinite list of populations, with the i-th element being the population of the i-th generation.

This library also provides some generic mutation function builders.

ITEA

itea :: NFData a => Mutation a -> Fitness a -> Population a -> Rnd [Population a]

Creates a stream of generations the i-th element corresponds to the population of the i-th generation.

initialPop

Arguments

:: NFData a

=> Int
maxTerms

-> Int
nPop

-> Rnd (Term a)
random term generator

-> Fitness a
fitness function

-> Rnd (Population a)

Generate an Initial Population at Random

tournament :: Population a -> Int -> Rnd (Population a)

Tournament Selection

given the concatenation of the previous population and the mutated children, it will return a sampled selection of these combined population with the same size as the original population.

chooseOne :: Population a -> Rnd (Solution a)

step :: NFData a => Mutation a -> Fitness a -> Int -> Population a -> Rnd (Population a)

Perform one iteration of ITEA

Parallel random functions

parRndMap :: NFData c => Int -> (a -> Rnd b) -> (b -> Maybe c) -> [a] -> Rnd [c]

Runs in parallel the composition of a function that generates random effects with a function that maybe returns a result.

numberOfSplits :: Int -> Int

Calculates the number of splits as twice the number of cores

genNseeds :: Int -> StdGen -> [StdGen]

Generates n random seeds.

genseeds :: StdGen -> [StdGen]

Generates an infinite list of random seeds.

parMaybeMap :: NFData b => Int -> (a -> Maybe b) -> [a] -> [b]

Runs a computation that may returns a result in parallel.