IT.Regression

Description

Definitions of IT data structure and support functions.

Documentation

type Vector = Vector Double

IT specific stuff

regSin :: Transformation Double

Transformation Functions

regCos :: Transformation Double

Transformation Functions

regTan :: Transformation Double

Transformation Functions

regTanh :: Transformation Double

Transformation Functions

regSqrt :: Transformation Double

Transformation Functions

regAbsSqrt :: Transformation Double

Transformation Functions

regLog :: Transformation Double

Transformation Functions

regExp :: Transformation Double

Transformation Functions

regTrig :: [Transformation Double]

regNonLinear :: [Transformation Double]

regLinear :: [Transformation a]

regAll :: [Transformation Double]

Error measures

mean :: Vector -> Double

Mean for a vector of doubles

var :: Vector -> Double

Variance for a vector of doubles

meanError

Arguments

:: (Double -> Double)
a function to be applied to the error terms (abs, square,...)

-> Vector
fitted values

-> Vector
target values

-> Double

generic mean error measure

Common error measures for regression:

mse :: Vector -> Vector -> Double

Mean Squared Error

mae :: Vector -> Vector -> Double

Mean Absolute Error

nmse :: Vector -> Vector -> Double

Normalized Mean Squared Error

rmse :: Vector -> Vector -> Double

Root of the Mean Squared Error

rSq :: Vector Double -> Vector Double -> Double

R^2

predict :: Matrix Double -> Vector -> Vector

Predict a linear model

data RegStats

Store regression statitstics

Constructors

RS

Fields

_rmse :: Double

_mae :: Double

_nmse :: Double

_r2 :: Double

_weights :: Vector

the fitted weights for the IT expression

Utility functions

isInvalid :: Double -> Bool

A value is invalid if it's wether NaN or Infinite

isValid :: [Double] -> Bool

a set of points is valid if none of its values are invalid and the maximum abosolute value is below 1e150 (to avoid overflow)

exprToMatrix :: Dataset Double -> Expr Double -> Matrix Double

evaluate an expression to a set of samples

(1 LA.|||) adds a bias dimension

cleanExpr :: Dataset Double -> Expr Double -> Expr Double

Clean the expression by removing the invalid teerms

notInfNan :: Solution Double RegStats -> Bool

Checks if the fitness of a solution is not Inf nor NaN.

parMapChunk :: Int -> (Expr Double -> Matrix Double) -> [Expr Double] -> [Matrix Double]

Parallel strategy for evaluating multiple expressions

fitnessReg :: Int -> Dataset Double -> Vector -> [Expr Double] -> Population Double RegStats

Fitness function for regression

Split the dataset into twice the available cores evaluate the expressions in parallel run a Linear regression on the evaluated expressions Remove from the population any expression that leads to NaNs or Infs

fitnessTest :: Dataset Double -> Vector -> Solution Double RegStats -> RegStats

Evaluates an expression into the test set. This is different from fitnessReg since it doesn't apply OLS.

roundoff :: RealFrac a => a -> a

Experimental: round off floating point to the 1e-10 place.

bogusSolution :: RegStats

buildSolution :: Vector -> Vector -> Vector -> RegStats

solveOLS :: Matrix Double -> Vector -> Vector

Solve the OLS *zss*w = ys*

regress :: Vector -> Expr Double -> Matrix Double -> Solution Double RegStats

Applies OLS and returns a Solution if the expression is invalid, it returns Infinity as a fitness

Orphan instances

IT Double
IT Instance for regression evals an expression to sum(w_i * term_i) + w0 with term_i = f_i(p_i(xs)) and p_i(xs) = prod(x^eij)

Instance details

Methods

itTimes :: Dataset Double -> Interaction -> Column Double

itAdd :: [Column Double] -> Column Double

itWeight :: Double -> Column Double -> Column Double