IT.Regression

Description

Definitions of IT data structure and support functions.

IT specific stuff

data Task

Constructors

Regression

Classification

ClassMult

type FitFun = Vector -> Vector -> Double

data Penalty

Constructors

NoPenalty

Len Double

Shape Double

predict :: Matrix Double -> Vector -> Vector

Predict a linear model

solveOLS :: Matrix Double -> Vector -> Vector

Solve the OLS *zss*w = ys*

isInvalidMatrix :: Matrix Double -> Bool

regress :: Matrix Double -> Vector -> [Vector]

Applies OLS and returns a Solution if the expression is invalid, it returns Infinity as a fitness

classify :: Matrix Double -> Vector -> [Vector]

classifyMult :: Matrix Double -> Vector -> [Vector]

fitTask :: Task -> Matrix Double -> Vector -> [Vector]

Fitness function for regression

Split the dataset into twice the available cores evaluate the expressions in parallel run a Linear regression on the evaluated expressions Remove from the population any expression that leads to NaNs or Infs it was fitnessReg

predictTask :: Task -> Matrix Double -> [Vector] -> Vector

evalPenalty :: Penalty -> Int -> Double -> Double

applyMeasures :: NonEmpty Measure -> Vector -> Vector -> [Double]

evalTrain :: Task -> NonEmpty Measure -> Constraint -> Penalty -> Dataset Double -> Vector -> Dataset Double -> Vector -> Expr -> Maybe Solution

evalTest :: Task -> NonEmpty Measure -> Dataset Double -> Vector -> Solution -> Maybe [Double]

Evaluates an expression into the test set. This is different from fitnessReg since it doesn't apply OLS. It was: fitnessTest