ITEA.Config

Description

Configuration parsing and report generation.

Documentation

class Monoid a => Valid a b | a -> b, b -> a where

Class of types that can be validate

Methods

validateConfig :: a -> b

data Param a

A parameter is either empty (None) or Has something

Constructors

None

Has a

fromParam :: Param a -> a

Extract parameter. This is a partial function.

data UncheckedMutationCfg

Unchecked mutation config

Constructors

UMCfg

Fields

_expLim :: Param (Int, Int)

_termLim :: Param (Int, Int)

_nzExp :: Param Int

_transFun :: Param [Transformation]

_measures :: Param [String]

data MutationCfg

Validated mutation config

Constructors

MCfg (Int, Int) (Int, Int) Int [Transformation] [String]

exponents :: Int -> Int -> UncheckedMutationCfg

Generates a configuration with only _expLim holding a value.

termLimit :: Int -> Int -> UncheckedMutationCfg

Generates a configuration with only _termLim holding a value.

nonzeroExps :: Int -> UncheckedMutationCfg

Generates a configuration with only _nzExp holding a value.

transFunctions :: [Transformation] -> UncheckedMutationCfg

Generates a configuration with only _transFun holding a value.

measures :: [String] -> UncheckedMutationCfg

Generates a configuration with only _measures holding a value.

getMaxTerms :: MutationCfg -> Int

getMeasure :: MutationCfg -> [Measure]

parseFile :: String -> (Matrix Double, Vector)

Parse a numerical csv file into predictors and target variables

withMutation :: MutationCfg -> Int -> (Mutation, Rnd Term)

Creates the mutation function and also returns the random term generator (for initialization)

Datasets configuration

data UncheckedDatasets

Constructors

UD

Fields

_trainset :: Param String

_testset :: Param String

data Datasets

Constructors

D String String

trainingset :: String -> UncheckedDatasets

sets the training and test data set names

testset :: String -> UncheckedDatasets

sets the training and test data set names